using int32_t for packet sample type

+++ src/packet.c
@@ -1,11 +1,11 @@
 #include "packet.h"
 
-const double gnuitar_sample_max = GNUITAR_SAMPLE_MAX;
+const int32_t gnuitar_sample_max = GNUITAR_SAMPLE_MAX;
 
-const double gnuitar_sample_min = GNUITAR_SAMPLE_MIN;
+const int32_t gnuitar_sample_min = GNUITAR_SAMPLE_MIN;
 
 void
-gnuitar_packet_mul(struct GnuitarPacket *packet, double n)
+gnuitar_packet_mul(struct GnuitarPacket *packet, int32_t n)
 {
     unsigned int i;
     for (i = 0; i < packet->len; i++) {
@@ -14,7 +14,7 @@ gnuitar_packet_mul(struct GnuitarPacket *packet, double n)
 }
 
 void
-gnuitar_packet_div(struct GnuitarPacket *packet, double n)
+gnuitar_packet_div(struct GnuitarPacket *packet, int32_t n)
 {
     unsigned int i;
     for (i = 0; i < packet->len; i++) {
